+++ tests/importable_module.py
@@ -1,0 +1,1 @@
+ignore_app_name_mismatch = True