+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Change log
 
 - Added a system check verifying that the appropriate ``unique_together`` value
   is set when using the ``LanguageAndTranslationOfMixin``.
+- Added a system check for the ``app_name`` value of application URLconf
+  modules.
 
 
 `3.3`_ (2022-03-03)